// web/projects/portal/src/app/composer/gui/vs/assembly-type.ts
/*
 * inetsoft-web - StyleBI is a business intelligence web application.
 * Copyright © 2024 InetSoft Technology (info@inetsoft.com)
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
export enum AssemblyType {
   TABLE_VIEW_ASSET = 101,
   CHART_ASSET = 102,
   CROSSTAB_ASSET = 103,
   FORMULA_TABLE_ASSET = 104,
   SLIDER_ASSET = 106,
   SPINNER_ASSET = 107,
   CHECKBOX_ASSET = 108,
   RADIOBUTTON_ASSET = 109,
   COMBOBOX_ASSET = 110,
   TEXT_ASSET = 111,
   IMAGE_ASSET = 112,
   GAUGE_ASSET = 113,
   THERMOMETER_ASSET = 114,
   SLIDING_SCALE_ASSET = 115,
   CYLINDER_ASSET = 116,
   SELECTION_LIST_ASSET = 117,
   SELECTION_TREE_ASSET = 118,
   TIME_SLIDER_ASSET = 119,
   CALENDAR_ASSET = 120,
   EMBEDDEDTABLE_VIEW_ASSET = 123,
   LINE_ASSET = 125,
   RECTANGLE_ASSET = 126,
   OVAL_ASSET = 127,
   GROUPCONTAINER_ASSET = 124,
   CURRENTSELECTION_ASSET = 128,
   TEXTINPUT_ASSET = 129,
   SUBMIT_ASSET = 130,
   UPLOAD_ASSET = 134,
   PAGEBREAK_ASSET = 135,
   ANNOTATION_ASSET = 131,
   ANNOTATION_LINE_ASSET = 132,
   ANNOTATION_RECTANGLE_ASSET = 133,
   VIEWSHEET_ASSET = 200
}